+++ app/src/pages/Dashboard/EventsOverviewPage/fragments/EventsPageHeader.test.tsx
@@ -21,14 +21,15 @@ describe('Events Page Header tests', () => {
   };
   test('page will not crash', async () => {
     const { store } = configureStore();
-    await render(<EventsPageHeader onFilterChange={onFilterChange} filters={filter} />, store);
-    expect(screen.getAllByRole('button')).toHaveLength(6);
+    await render(<EventsPageHeader entries={[]} onFilterChange={onFilterChange} filters={filter} />, store);
+    // Two filter buttons along with the create meeting link that has a role=button
+    expect(screen.getAllByRole('button')).toHaveLength(3);
     expect(screen.getAllByRole('link')).toHaveLength(1);
   });
 
   test('click on filter will call onFilterChange function', async () => {
     const { store } = configureStore();
-    await render(<EventsPageHeader onFilterChange={onFilterChange} filters={filter} />, store);
+    await render(<EventsPageHeader entries={[]} onFilterChange={onFilterChange} filters={filter} />, store);
     const favoriteMeeting = screen.getByTestId('favoriteMeeting');
     expect(favoriteMeeting).toBeInTheDocument();
     fireEvent.click(favoriteMeeting);